Enforce table integrity with transforms and move non-React plugin code to @lexical/table
