+++ packages/lexical-playground/src/plugins/TablePlugin.tsx
@@ -8,22 +8,13 @@
 
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 import {
-  $createTableNodeWithDimensions,
   INSERT_TABLE_COMMAND,
+  TableCellNode,
   TableNode,
+  TableRowNode,
 } from '@lexical/table';
-import {
-  $insertNodes,
-  COMMAND_PRIORITY_EDITOR,
-  createCommand,
-  EditorThemeClasses,
-  Klass,
-  LexicalCommand,
-  LexicalEditor,
-  LexicalNode,
-} from 'lexical';
+import {EditorThemeClasses, Klass, LexicalEditor, LexicalNode} from 'lexical';
 import {createContext, useContext, useEffect, useMemo, useState} from 'react';
-import * as React from 'react';
 import invariant from 'shared/invariant';
 
 import Button from '../ui/Button';
@@ -52,9 +43,6 @@ export type CellEditorConfig = Readonly<{
   readOnly?: boolean;
   theme?: EditorThemeClasses;
 }>;
-
-export const INSERT_NEW_TABLE_COMMAND: LexicalCommand<InsertTableCommandPayload> =
-  createCommand('INSERT_NEW_TABLE_COMMAND');
 
 export const CellContext = createContext<CellContextShape>({
   cellEditorConfig: null,
@@ -155,28 +143,16 @@ export function TablePlugin({
 }): JSX.Element | null {
   const [editor] = useLexicalComposerContext();
   const cellContext = useContext(CellContext);
-
   useEffect(() => {
-    if (!editor.hasNodes([TableNode])) {
-      invariant(false, 'TablePlugin: TableNode is not registered on editor');
+    if (!editor.hasNodes([TableNode, TableRowNode, TableCellNode])) {
+      invariant(
+        false,
+        'TablePlugin: TableNode, TableRowNode, or TableCellNode is not registered on editor',
+      );
     }
-
+  }, [editor]);
+  useEffect(() => {
     cellContext.set(cellEditorConfig, children);
-
-    return editor.registerCommand<InsertTableCommandPayload>(
-      INSERT_NEW_TABLE_COMMAND,
-      ({columns, rows, includeHeaders}) => {
-        const tableNode = $createTableNodeWithDimensions(
-          Number(rows),
-          Number(columns),
-          includeHeaders,
-        );
-        $insertNodes([tableNode]);
-        return true;
-      },
-      COMMAND_PRIORITY_EDITOR,
-    );
-  }, [cellContext, cellEditorConfig, children, editor]);
-
+  }, [cellContext, cellEditorConfig, children]);
   return null;
 }
